+++ Services/IRouletteService.s.cs
@@ -2,6 +2,6 @@ namespace RuletaOnline.Services
 {
     public interface IRouletteService
     {
-        int CreateRoulette();
+        long CreateRoulette();
     }
 }
